+++ src/Writing.jsx
@@ -1,0 +1,7 @@
+export function Writing() {
+  return (
+    <div>
+      <iframe src="https://apack.bairui.dev" style={{width: "100vw", height: "100vh", border: "none"}} />
+    </div>
+  );
+}
